tricking webapp to process cv data like its manual dart data
Considering I broke the regular simulation in my CV integration attempts, i'm gonna have to revert this commit and a couple others. I just want this here so I have a record of attempting to trick the webapp to process the dart data the same way. God give me strength to fix this bug